requirements.txt
Django==1.6.2
Markdown==2.4
Pillow==2.3.1
South==0.8.4
argparse==1.2.1
django-filter==0.7
djangorestframework==2.3.13
easy-thumbnails==1.5
pygeoip==0.3.1
pytz==2014.2
wsgiref==0.1.2
django-countries==1.5
